requirements.txt
geopandas==0.12.1
numpy==1.21.6
pandas==1.4.2
Pillow==9.3.0
plotly==5.9.0
Shapely==1.8.5.post1
streamlit==1.11.0
